Fixed comment and change int to float.

+++ FIXED_BATCH_4/maximum_number.py
@@ -1,20 +1,20 @@
-#Create a list
+# Create a list
 numbers_list = []
 
-#Use for loop to allow users to input 10 numbers
+# Use for loop to allow users to input 10 numbers
 while True:
     try:
-        number = int(input("Enter the number: "))
+        number = float(input("Enter the number: "))
     
-        #Update the list
+        # Update the list
         numbers_list.append(number)
 
     except ValueError:
         print("Invalid")
         break
 
-#Check which is the highest number
+# Check which is the highest number
 max_number = max(numbers_list)
         
-#Print the result
+# Print the result
 print(f"Result: {max_number}")
